destructure control-system to move to modules
