+++ firmware/projects/FrontController/accumulator/accumulator.cc
@@ -1,8 +1,6 @@
-#include "control-system/battery_monitor.hpp"
+#include "accumulator.hpp"
 
 #include <optional>
-
-#include "control-system/enums.hpp"
 
 BatteryMonitor::BatteryMonitor() : current_status_(std::nullopt) {}
