+++ firmware/projects/FrontController/accumulator/accumulator.hpp
@@ -2,7 +2,7 @@
 
 #include <optional>
 
-#include "enums.hpp"
+#include "../enums.hpp"
 
 struct ContactorFeedback {
     enum class State : bool {
