--- firmware/projects/FrontController/control-system/.gitignore
@@ -1,2 +1,0 @@
-*.o
-*.exe
