--- firmware/projects/FrontController/control-system/CMakeLists.txt
@@ -1,11 +1,0 @@
-add_library(control-system INTERFACE)
-add_subdirectory(src)
-target_include_directories(control-system INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
-add_dependencies(control-system generated_can)
-
-if(TESTING_ENABLED)
-    target_link_libraries(tests PRIVATE control-system)
-
-    file(GLOB CONTROL_SYS_TEST_SRC "test/test*.cc")
-    target_sources(tests PRIVATE ${CONTROL_SYS_TEST_SRC})
-endif()
