--- firmware/projects/FrontController/control-system/src/motor_interface.cc
@@ -1,2 +1,0 @@
-// #include "control-system/enums.hpp"
-// #include "control-system/motor_interface.hpp"
