+++ firmware/projects/FrontController/enums.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../generated/can/veh_messages.hpp"
+#include "generated/can/veh_messages.hpp"
 
 using DiSts = generated::can::TxFC_Status::DiStatus_t;
 using GovSts = generated::can::TxFC_Status::GovStatus_t;
