+++ firmware/projects/FrontController/governor/governor.hpp
@@ -2,7 +2,7 @@
 
 #include <optional>
 
-#include "enums.hpp"
+#include "../enums.hpp"
 
 class Governor {
 public:
