+++ firmware/projects/FrontController/vehicle_dynamics/test_vehicle_dynamics_calc.cc
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "control-system/vehicle_dynamics_calc.hpp"
+#include "vehicle_dynamics_calc.hpp"
 
 using namespace ctrl;
