+++ firmware/projects/FrontController/vehicle_dynamics/vehicle_dynamics_calc.cc
@@ -1,4 +1,4 @@
-#include "control-system/vehicle_dynamics_calc.hpp"
+#include "vehicle_dynamics_calc.hpp"
 
 #include <algorithm>
